backend/requirements.txt
# Core
fastapi==0.110.1
uvicorn==0.25.0
python-dotenv==1.2.1
pydantic==2.12.4

# Database
motor==3.3.1
pymongo==4.5.0

# OpenAI
openai==1.99.9

# Fault tolerance
tenacity==9.1.2

# CORS
starlette==0.37.2
